jmap_util.c: add KEEP_EXISTING flag to JMAP patchobject_apply
A patch object that patches an existing JSON key replaces the value
of that key with whatever value is set in the patch. But a patch
object coming from an untrusted source, such as the JSPROP property
in iCalendar data, should not be able to replace existing values.

This patch adds the KEEP_EXISTING flag to allow for this. It instructs
the patching function to only apply patches for JSON keys which do
not exist in the patched JSON object.