jmap_util.c: use parser-internal buffer in jmap_parser_path
This refactors jmap_parser_path to use the parser-internal
buffer for the return value, rather than encoding the path
in a buffer passed as a function argument. As a result,
the memory pointed to by the return value only is valid
until the next call to any of the parser functions.

This is less brittle than it sounds. Most of the calls to
jmap_parser_path already passed in the parser's internal
buffer as the function argument, and for the remaining calls
we now just set the caller-supplied buffer to the return
value of the jmap_parser_path function.